Simplify the implementation of chpl__first|lastForIter for bounded ranges
Looking at the 'anonymousRangeIter' failure caused by merging #21332,
it appeared the the expression for computing the first/last element in
a typical bounded range iteration had become more complicated by my
refactoring.  Here, I'm splitting the common case of a completely
bounded range into its own case and handling it identically to how we
did prior to #21332, which should limit the complicated cases to
iterations over unbounded ranges of enum and bool.  This resolves the
anonymousRangeIter test and makes the generated code much more similar
to what we had before.

This change seems more likely to affect/improve the performance of
range iteration than #21405, so I am cautiously optimistic that it
will restore the performance of cases that slipped as tracked in
https://github.com/Cray/chapel-private/issues/4228

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>