Fix failure to forward error code from gpu codegen
When GPU codegen has a failure, it exits the GPU subprocess with a non-zero
exit code. The main compiler process then sees that the code from 'wait' is
non-zero and also exits with that code. But the code from 'wait' is not just
the code the subprocess exited with, so this ends up being equivalent to
'exit(0)'. Extract the actual exit code from the code set by 'wait' and exit
with that code instead.

Add a test that causes GPU codegen to fail. start_test was previously failing
for this test with errors like:
"[Error could not locate executable gpuCodegenErrorExitStatus for gpu/native/environment/gpuCodegenErrorExitStatus]".

Now it correctly detects that the compilation failed and stops there and
diffs the compilation output vs. the .good file.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>