Update spack version for mason --external
When running `mason external --setup` on Ubuntu 22.10 or SLES 15,
I was seeing an error from Spack itself. That is because we had pinned
version v0.15.4 in the mason source code and this version is probably
too old to work with Python 3.10.

The error included this from within Spack code:
  AttributeError: module 'collections' has no attribute 'MutableMapping'

Once I updated the version number to the current Spack release
(v0.19.0), `mason external --setup` still did not work, because of
differences in how `spack --version` output changed. So, I updated the
trimming used when processing the output of `spack --version`.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>