Resolve forwarding statements for use in resolving calls
Note, it's not so easy for resolving calls to track when forwarding is
used because there could be multiple layers of forwarding involved
(e.g., a record could forward to another record, which forwards again).
This information will have to be figured out again when the resolved
code is lowered.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>