Remove unnecessary and incorrect assertion when figuring out the init phase
This part of the code was copied from the checking in normalize, but the AST
will have changed in between those passes, so it's perfectly reasonable for a
forall to have the fRecIterIRdef field set.  However, it shouldn't be relevant
when determining the initializer's phase at this time, as the normalize checking
of initializers should have errored if there was any `super.init()` or
`this.init()` call in code that would be included there.

This should fix the internal assertion error when compiling CHAMPS from last
night - it fixes a small test program I made from that, but I'll also check
CHAMPS itself.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>