Fix a few more tests and add a `throws` tag to a bigint cast operator
Opts into using the new throwing behavior for all these tests.  As a result,
discovered that one of the cast operator implementations calls the throwing
initializer now, so it needs to either be marked throws itself or have a `try!`
in its body.  Decided to make it throw, since there is precedent for other cast
operators throwing.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>