Remove the link to the errorCode type in the initializer deprecation message
We needed the `~OS` prefix to point to the type correctly, but having the `~`
broke the elision of `:type:` prefixes for the deprecation message.  The link
seems like the least important part of that, so just remove it instead of
fussing with it further.

I could alternatively have removed the `~` portion and let the docs display have
the full path to the type, but it didn't seem worthwhile to wrestle with that
further at the moment.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>