Keep track of forwarding actual type through disambiguation
This enables the dyno resolver to potentially disambiguate
in a case like the below where it is necessary to associate
the candidate with the actual reciever type used after forwarding
in order to choose between the options.

    record A {
      forwarding var b : B;
      forwarding var c : C;
    }

    proc C.foo() { }
    iter B.these() : C { }
    var a : A;
    a.foo(); // through a.b, or through a.c?

It might be that this particular case needs to be an error in the
future. Nonetheless it is good for the dyno resolver's components to be
able to handle it.

The approach is to use a separate vector to track this information when
forwarding is used. That will hopefully reduce the performance impact of
tracking this additional information (since we expect forwarding is
rarely used). Additionally, it makes the reason for tracking this
additional information is clearer in the compiler source code (at least
compared to making the candidates vector store a tuple).

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>