Upgrade spack version used by mason
We've been using spack 0.15.4 for mason-external for a while because
newer versions didn't work when spack was in a path longer than ~100
chars. However, not upgrading has left us with a spack that doesn't work
on M1 macs or python >= 3.8, which seems worse to me.

This upgrades spack to 0.19.0, which works on M1 and supports python
2.7/3.6-3.11. It also updates the mason-external tests to use a
`MASON_HOME`/`SPACK_ROOT` in `CHPL_TEST_TMP_DIR`, which generally has a
short directory structure so should avoid the path length issues. I'm
using `CHPL_TEST_TMP_DIR` instead of just a new tmpdir because we take
extra care to clean up `CHPL_TEST_TMP_DIR`.

A side benefit of this is that that dir is usually in `/tmp` so on some
systems we avoid cloning spack into NFS/lustre so this should save some
time and reduce load on shared filesystems.

Resolves 19517
Resolves Cray/chapel-private 3206
Resolves Cray/chapel-private 3518
Part of Cray/chapel-private 4110

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>