Propagate user labels when we are copying loop statements
In #21453 it was pointed out that we weren't finding a loop label
within the context of a 'coforall' loop.  It turned out that the
reason for this was that when we were copying the loops for the
coforall loops, their labels were not being preserved.

In this PR, I've added copies of the labels to the copyInner()
routines of BlockStmt, ForLoop, WhileDoStmt, and WhileDoStmt.  It
seems like just the first should be sufficient since the others are
sub-classes of it, yet... it doesn't seem to be.  So this PR fixes the
issue, but in an unsatisfying way.  I'll hope for a reviewer who knows
what I'm doing wrong and can help me improve it before merging.

Added a test that didn't work before, but does now.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>