Make primary method `this` formals not have type exprs.
Prior to this commit, Dyno inserted an implicit `this` formal for all
methods. For secondary methods, the type expression of this implicit
formal was derived from the type expression in the user code. That is to
say, the following secondary method:

```Chapel
proc (someTypeExpr).foo() {}
```

translated to the following:

```Chapel
proc foo(this: someTypeExpr) {}
```

In the case of primary methods, however, the type expression was simply
the name of the class in which the method is being declared. For
instance, the following:

```Chapel
record R {
    proc foo() {}
}
```

becomes:

```Chapel
proc foo(this: R) {}
```

However, this led to problems in the case that a field of the record had
the same name as the record itself. In these situations, the Dyno
resolver would decide that `this` refers not to the record, but to the
field. This commit/PR changes the behavior to instead translate the
function as follows:

```Chapel
proc foo(this) {}
```

The resolver is also adjusted to provide special treatment to `this`
formals without a type expression; the ID that the type expression
refers to is automatically assumed to be that of the record or class in
which the primary method is being defined.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>