Deprecate the 'sameFile' overload with 'file' arguments
The 'sameFile' function that takes 'file' arguments is the only function
in the 'FileSystem' module that takes 'file' arguments. Deprecate it for the
sake of consistency with the rest of the functions in 'FileSystem'.

Update tests to use the 'string' version of 'sameFile' where 'file.path' is
available. Stop testing on a few tests that were calling 'sameFile' on an
unset 'file'.

Add a deprecated test for 'sameFile'.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>