Squash AST-build-time errors about breaks being in illegal scopes
As demonstrated by labeledLoopWithinNestedCoforall.chpl on 'main'
prior to this PR (itself a variation on the test reported in
https://github.com/chapel-lang/chapel/issues/21453#issuecomment-1413742819),
our build-time checkControlFlow() check was issuing errors for labeled
breaks that are actually legal.

In the time I had to look into this, I wasn't able to determine what
about its current logic was wrong, other than to see that something
about the loopSet and labelSet are being maintained in a way that
prevents the labeled loop from having its label added to the set.  I
tried to understand the logic, but couldn't quickly grok what loopSet
is meant to be representing at any given time.

Anyway, the net result was that it caused the checks to think that the
break was breaking out of an outer on-statement or coforall-loop
rather than the labeled loop in question.  My guess is that the logic
isn't understanding nesting and/or labeled break statements correctly.

Since I couldn't get to the bottom of it quickly, I just squashed the
error in the case that it's a labeled break, since I consider it worse
to issue errors for correct programs than to fail to issue errors for
incorrect programs (not that I have such an incorrect program in
mind...  specifically, the only test in our suite that checks this
logic — test/parallel/begin/deitz/test_control_flow.chpl — is still
working after these changes).  I'm also reasonably optimistic that in
cases where we truly have bad control flow, it will lead to errors
later in compilation.  For example, it's hard to imagine that a
'break;' that jumps out of an on-statement or coforall will find its
labeled loop after the compiler has done its outlining to implement
those features, and the code later in the compiler that I was
inspecting when working on #21476 was definitely more cognizant of
scoping.

While here, I removed a compiler message "the following error is a
current limitation" because it felt like a weird message, and a sloppy
implementation.  I'm also not sure how much value it adds, and am not
seeing a sign, with a quick read of the spec, that we've made any
promises to support breaks out of on-statements.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>