Add a test locking in that we throw a SystemError when the c_FILE is closed
Unfortunately the message stored in the SystemError is not particularly clear in
that case (it's 'Function not implemented (in init with path "unknown")', which
we get from the call to `fileno` in `qio_file_init`.  This error might vary by
system, the man page for fileno only lists EBADF as an error code).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>