Change USR_FATAL_CONT() into USR_FATAL() since only bad things will happen...
...when this error is triggered, as Vass pointed out in his review.
And a user can always use --ignore-errors if they want to see those
bad things firsthand.

I revamped my test a bit based on that and added a new case that I
think we ought to support similarly, but don't, as a future.  I
also re-worded the error message with fresh eyes.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>