Skip this test under valgrind
Valgrind complains about trying to access memory that has been free'd, but
that's a case we're intentionally trying to test, so it's okay not to worry
about it.

I was seeing a segfault with valgrind on Mac - if that proves to be symptomatic
of a broader problem, we should remove the test.  Leave a note on the test
indicating it is okay to remove it.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>