Use a different visited set per forwarding target
As this code stands now, the candidates are gathered and filtered in a
loop for each forwarding target. So, if a class is visited for one of
these targets but there were no candidates, that class won't be visited
again for the next target, even if it would result in a candidate.

An additional complication is that, because of method visibility rules,
the scopes consulted to find visible functions depend on the method
receiver type.

So, this commit adjusts the forwarding logic to use a different visited
set per 'forwarding' target.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>