Add a 'metadata' argument to 'copyTree'
If the 'metadata' argument is true, copy the UID and GID from the original
directory.

When copying files, forward the 'metadata' argument to 'copy'.

Add a test that tries to copy a directory owned by root with metadata=true. Since
that's not allowed as a non-root user it expects and gets a Permission error.

I also tested a modified version of this test that prints the UID and GID of the
copied directory and ran it with 'start_test --launchcmd fakeroot' and verified
that the UID and GID of the copy were both 0 (root).

Signed-off-by: David Iten <daviditen@users.noreply.github.com>