Adjust .good for negative stride #21496
The test guru-interface-example.chpl creates a temp negative-stride array
when calling reshape(), in order to reverse the order of elements
of a unit-stride array and store the result in another unit-stride array.

This PR adjusts its .good to account for the resultingwarning,
the same way I adjusted other tests in #21496.

Trivial, not reviewed.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>