Prototype implementation of Encoders/Decoders (formatters)
This commit introduces new fields to the _channel type to support a
"formatter", which is responsible for implementing the format used to
read and write values. The ``_fmt`` field is constrained by ``fmtType``, and
is accessed with the new paren-less ``formatter`` method. This formatter
can currently be set with the ``withFormatter`` method, which will
return an alias to the original channel which will point to the same
position in the channel, but will now read through the desired
formatter. This support can be enabled or disabled with the
``--[no-]use-io-formatters`` flag.

A prototype of the "default" writer and reader is included in this
commit. Channels will invoke the ``encode`` method when writing, and the
``decode`` method when reading. These default implementations maintain
internal state in order to print commas in the correct locations for
fields, and also to correctly "flatten" inherited classes.

FormatWriters have the option of invoking a user-defined ``encodeTo`` method.
The compiler is expected to provide a compiler-generated ``encodeTo``
method that invokes the ``writeTypeStart``, ``writeField``, and
``writeTypeEnd``  methods on the formatter. It is useful for a
FormatWriter to maintain some kind of internal state to track its
position within an object, and so it is generally useful to invoke
``encodeTo`` with a newly-constructed version of the given FormatWriter
to act as a "clean slate".

FormatReaders have the option of invoking an initializer for the desired
type, and passing a ``fileReader`` argument. As with FormatWriters, it
is useful to invoke such initializers with a newly-constructed
FormatReader with fresh internal state. The compiler-generated
initializer in such cases is expected to invoke ``readTypeStart``,
``readField``, and ``readTypeEnd``. FormatReaders also have the option
of invoking a ``decodeFrom`` type method for types that don't support
initializers.

When attempting to read an ``owned`` or ``shared`` class type, the
current implementation will instead request the FormatReader to read an
unmanaged version of the class type. This will hopefully allow for
simpler reasoning about genericity for FormatReader authors. The
``fileReader`` methods that invoke the FormatReader will be expected to
turn the ``unmanaged`` result into the correct managed type on the way
out.

At present, special handling of ioLiteral, ioNewline, ioChar, and ioBits
is requried as FormatReaders do not accept values.

This commit also modifies ``file.readerHelper``, ``file.linesHelper``,
and ``file.writerHelper`` to simplify some of the code involving the
creation of a ``fileReader`` or ``fileWriter``.

The implementation of ``_channel.read(type t)`` has been updated to
allow for the reading of non-default-initializable types when formatters
are enabled.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>