Formatters support in ChapelIO
Adds ``encodeToDefaultImpl`` for the compiler to invoke when generating
the default ``encodeTo`` method for types.

Also provides ``chpl__isFileReader`` for use in the where-clauses of
compiler-generated initializers that are meant to accept a
``fileReader``. We currently must leave the argument fully-generic
because ``fileReader`` is not available in-scope without pulling in the
entire IO module.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>