Support for compiler-generated initializer that takes a fileReader
This commit adds a new method, ``buildReaderInitializer``, to
AggregateType which will create a new compiler-generated initializer
that accepts a ``fileReader`` to be used when reading in a type from a
file. Technically the argument to this initializer will be fully-generic
because ``fileReader`` isn't available without explicitly use-ing the
``IO`` module. The ``chpl__isFileReader`` function is available from the
ChapelIO module and is used in a where-clause to constrain the
initializer and avoid ambiguity.

Generic types have ``type`` and ``param`` formals along with the
``fileReader`` argument. It is expected that such initializers will
either be invoked with those arguments explicitly, or with a type-alias
(which leans on the compiler to insert those arguments).

Currently these "reader initializers" are not generated for internal
modules because we cannot adequately constrain the types or use a
where-clause.

A new flag, FLAG_DEFAULT_INIT, is added to help distinguish between the
two initializers.

This new initializer will invoke ``readTypeStart`` on the given
formatter, and will then invoke ``super.init(...)`` with the
``fileReader`` argument in the case of classes. Then, each field will be
initialized with a call to ``readField`` on the formatter. Finally,
``readTypeEnd`` is invoked.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>