Update modules/noakes/mod-init-works
This test was added to show an internal error.
It used a pattern like

   use ChapelStandard.M1;

to request the top-level scope.

However, that pattern is not documented anywhere, and does not work with
the dyno scope resolver. This PR adjusts it to use

   import M1;

instead, to refer to the M1 in the top-level scope.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>