Update comment as suggested by Andy; wordsmith/re-format deprecation warning
Andy pointed out that I copied the comment in arrayFind.chpl from
arrayFuncs.chpl without updating it for this context, so I've done
that here.

In looking at this, I also gave the deprecation warning itself another
look and realized that it could be improved to mention the new
two-argument array find, and to remove some markdown-style formatting
that we tend not to use in error messages.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>