Add InitResolver support for conditionals
Adds two public methods to InitResolver, 'fork' and 'merge', to help
apply initializer semantics to conditionals.

The 'fork' method simply returns a copy of the invoked InitResolver.

The 'merge' method, when given two InitResolvers, attemps to reconcile
any differences between the two and incorporate the unified state into
the invoked InitResolver. In particular, it will attempt to resolve
fields in one branch if the other branch was further initialized. Also,
the 'merge' method checks whether the two branches instantiated
different generic types.

If 'merge' is only supplied with one InitResolver, then it will copy
that InitResolver's state. This can occur when a then-statement has no
else-block.

This commit also updates 'thisCompleteIds_' to be a vector, so that we
can eventually track multiple occurrences of "this.complete()" for
better errors.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>