Fix default and const argument intents on tuples
Resolves #9104

Prior to this change, `f(arg: 2*int)` and `f(const arg: 2*int)` would
accept its actual argument by (const) reference. This was the not the
case, i.e., the rules were implemented correctly, when the tuple contained
at least one component whose type had a ref or const-ref default
argument intent or when the types of the actual and the formal differed.

The rules are defined at this link and in the subsection that follows it:
  https://chapel-lang.org/docs/1.29/language/spec/tuples.html#tuple-argument-intents

DETAILS

The compiler now inserts a cast from a tuple type to the same tuple type
in the affected cases, that is, when passing a tuple by default or 'const'
intent to a formal of the same type.

I implemented aggressive pruning to avoid inserting such a cast, for
example, when the actual argument is itself a default-intent formal or
a tuple literal that is not used elsewhere or when the invoked
function is already a tuple cast or an autocopy or initcopy.
Without such pruning the cast shows up in some places where the compiler
does not expect it, resulting in compilation failures or errors.

I special-cased the implementation of such a cast for tuples that are POD,
to simplify the generated code from field-wise copies to a single "move".

When creating a ContextCallExpr, I ensured that it is placed after the LAST

(in the AST order) of the CallExprs that it replaces. This is needed
when the original CallExpr appears at the statement level and requires
tuple cast(s). I observed this ifor the standalone expression `A[nextIdx];`
in: test/optimizations/sungeun/RADOpt/access2d.chpl

While there, I beefed up INT_ASSERTs in loopDetails.cpp with an AST argument
for source file+line info.

TEST CHANGES

* `accessChecks.chpl` : I added this variant of existing tests in order
to turn on bounds checking. I observed a bounds violation due to the above
issue with ContextCallExpr, which went unnoticed by the existing tests
because they threw --fast. I chose the 3-d mode as it was more complex.

* `const-checking-loop-error.bad` : the actual argument to modit() is
no longer affected by the (incorrectly allowed) changes to the formal
argument because the latter is now a copy of the former. Anecdotally,
if I change the intent of the modit() formal to `const ref`, writeln()
still writes the unmodified tuple. I believe this is because the copy
propagator knows that a value will not be changed when passed to
a function by a const-ref intent.

* `default-intents.chpl` : this is the former future test for #9104.
I beefed it up.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>