Do a configure-like check for MLI skipifs
Multi-locale interop (MLI) requires ZMQ. Currently, it requires a true
system install of ZMQ (we only provide `-lzmq` and don't have a way to
provide non-default `-L`/`-I` paths.) Previously, the skipif was just
checking for `libzmq.so` and assuming the headers would be available if
the shared object was found, but this isn't true on SLES15 and some
other systems.

For other ZMQ tests, we switched to using `pkg-config` for the skipifs
and providing lib/include paths, but since we don't have a way to
provide lib/include paths for MLI support that won't work here. Instead
this does a "configure-like" check to compile a small zmq test program
and skip based on whether that succeeded.

This isn't ideal, but is better than letting the tests fail on systems
with libzmq.so, but without headers.

Part of Cray/chapel-private 4292

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>