Serialize UniqueStrings using unique ids
Adds a notion of a 'string cache' to the Serializer and Deserializer
types. Instead of writing out the full string data for every occurrence
of a UniqueString, cache the results and only write out the unique
integer identifier. This string cache will be serialized and
deserialized separately and before occurrences of UniqueStrings in a
serialized file. When deserializing the unique id can be used to look up
the cached string in the Deserializer.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>