Serialize the entirety of BuilderResult
This commit adds support for serializing and deserializing all relevant
fields in a BuilderResult. A textual outline of the binary format has
also been included. As a performance optimization, the 'idToLocation_'
map is serialized and deserialized in such a way that the file path only
occurs once, rather than for every Location.

Other changes are as follows:
- remove 'magic' and 'version' numbers; library will handle such things
- add start/end sentinel strings for verification
- update methods to accept Serializer and Deserializer
- compare method is updated to handle new fields

Future performance TODOs:
- can try variable-byte encoding integers
- could try not loading Location data until needed
- can try recomputing the uAST IDs rather than storing them

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>