Add LibraryFile to chpl::parsing
This commit adds a new "LibraryFile" type to chpl::parsing. This type
reads in the header information and string cache from a given library
file.

The library file format is as follows (whitespace not significant):

```
<magic number, uint64_t>
<library version, major, int>
<library version, minor, int>
<chpl version, major, int>
<chpl version, minor, int>
<chpl version, update, int>

<user/std module descriptor, std::string>

N:<number of BuilderResult entries, uint64_t>
  0..N-1: <file path i, std::string><offset i, uint64_t>

M:<string cache size, uint64_t>
  0..M-1: <id i, int><string length, uint32_t><string, const char*>

0..N-1: <BuilderResult for file path i, BuilderResult>
```

The offsets stored in this format represent the offset from the start of
BuilderResult entries, rather than from the start of the file itself.

The ``LibraryFile::generate`` method accepts a list of paths to
serialize, a name for the library file, and a boolean to indicate
whether the serialized files represent user modules (rather than
internal or standard modules).

When reading a LibraryFile, only the header and string cache are read.
The stored offsets and this string cache are used later by the
``loadBuilderResultFromFile`` function to load an individual
BuilderResult. With this functionality ``parseFileToBuilderResult`` has
been updated to try reading from a LibraryFile if the desired '.chpl'
file has an associated library, rather than parsing the file.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>