Update compiler driver to support .dyno files
This commit updates the production compiler to generate ".dyno" files in
the frontend's ``LibraryFile`` format and to handle ".dyno" files on the
command line.

The ``--dyno-gen-lib`` flag is used to specify a ".chpl" file to store
in a library, and the file's basename (without the .chpl extension) is
used as the name for the ".dyno" file. The special string "<standard>"
can be used to store the internal and standard modules in a ".dyno" file.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>