Add tests for .dyno files
The new 'testGenLib' test is the main focus of this commit. The .precomp
for 'testGenLib' will generate .dyno files for MyMod, OtherMod, the
standard library, and testGenLib itself. 'testGenLib' will then try
using these .dyno files in the following ways:
- MyMod.dyno alone
- OtherMod.dyno alone
- both MyMod.dyno and OtherMod.dyno
- both user modules and the standard library as .dyno files

The 'buildAll' test uses all four .dyno files in a single 'chpl'
command, since we cannot replace 'testGenLib.chpl' for 'testGenLib'
itself.

The prediff for these tests relies on the debug trace output for dyno
queries, and determines success by the lack of fileTextQuery lines in
the output (which indicate usage of the parser).

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>