Have paratest.server test futures by default and add a -no-futures flag
This changes paratest.server to test futures by default in hopes of
cutting down on the number of PRs that go in cleanly except for some
.bad mismatches on futures.  Though this adds testing time by default
on legacy systems, for modern/fast ones, the additional overhead
should be negligible.

Also added here is a -no-futures flag that permits someone who's not
ready to run futures or wants to shave a bit of time off their runs to
opt back out of testing them by default.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>