CHPL_RT_COMM_OFI_EP_COUNT overrides maximum number of endpoints
The CHPL_RT_COMM_OFI_EP_COUNT environment variable overrides the maximum
number of endpoints reported by the ofi provider in the ep_cnt field of the
domain attributes. Some providers report an incorrect maximum that is too low
to allow one endpoint per thread. As a result, endpoints are not bound to
threads and the runtime cannot make use of endpoint ordering properties.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>