Update .bad file to include new warning
I'm not sure why this line of this test is generating this warning...
It doesn't seem like it should be.  Best guess I have is that when you
hit a syntax error in bison, the results are somewhat unpredictable
(?).  Not completely satisfying, but I don't feel like I know how I'd
deal with it otherwise, and don't feel as worried about it as I would
if this weren't a future.  Too bad the warning doesn't appear after
the first syntax errors, though...

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>