Deprecate and rename the Barriers module and Barrier type
The 'Barriers' module is deprecated and renamed to 'Collectives' and the
'Barrier' type is deprecated and renamed to 'barrier'.

Add a deprecated test that generates the two new deprecation warnings.

Add Collections to the module documentation.

Now that the type is named 'barrier' the 'barrier.barrier()' function looks
like a constructor. The compiler had an error saying to use initializers
instead of constructors for methods named the same as their parent type.
The error has been around for long enough that nobody should have constructors
anymore, so remove the error and allow it as a regular method.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>