Fix up some failures from #21633 and #21733
(watch out for those #21x33 PRs, kids!)

This fixes a few futures that I must've missed in my last paratest run
(prior to paratest running futures by default!) as well as an obvious
problem with the test of the bash autocompletion which I'm not sure
how it slipped past me.

It also fixes an issue from Ben McDonald's #21733 which I happened to
notice since he's likely out for the evening.

While here, I noticed that an erroneous condition that I'd noted
earlier in #21652 has been resolved, probably by #21614 as Michael
predicted, so I added a compopts line to lock that behavior in (and
removed another that, in retrospect, felt overly paranoid).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>