Yield tuples by value when appropriate
When yielding a tuple within an iterator with the default return intent,
yield the components by value instead of by reference. This is to match
the yield mode of yielding a component outside of a tuple, which is done
by value. For example, this makes, `yield (myRecord,1);` yield myRecord
by value just like `yield myRecord;` already does.

Iterators that wish to opt out, i.e., use the prior semantics,
can use FLAG_DONT_UNREF_FOR_YIELDS / pragma "do not unref for yields".
We plan to replace this pragma with tuple per-component-refness specifiers,
see #21717. Use this pragma in a couple of tests that need it.

The standalone and follower iterators for loop expressions use the prior
semantics. This is to allow the following loop body to update array elements,
see also `test/expressions/loop-expr/forall-over-zip-over-arrays.chpl`

    forall xy in ([ab in zip(A,B)] ab) do
      xy = (0,0);

Implementation-wise, the compiler inserts a call to initCopy() on the tuple
being yielded, which creates copies of the components.

While there, exit early from insertUnrefForArrayOrTupleReturn() when
the return type is known to be neither an array nor a tuple,
to avoid doing unnecessary work.
When the function is an iterator, process the yields then return.

The change in implementForallIntents.cpp converts an INT_ASSERT
to a USR_FATAL.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>