Add checking of yields for 'const ref' iterators
This follows a suggestion by the reviewer of this PR.
While there, I improved the checks for yields of 'ref' iterators.

The iterator 'sortedMap.items()' had an inappropriate 'const ref'
return intent, claiming to yield keys and values by reference.
That was not necessarily the case and in any case was not
helped by the return intent. I adjusted it to match 'map.items()'.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>