Track 'errorCollectionRoot' on dependency edges.
This is necessary because a dependency can be used not as an error
collection root by one query, but then be used as an error collection
root for another query. Whether or not errors should be silenced should
hinge on how a dependency was used by a query, and since it can be used
differently many times, that information should be stored per-use (i.e.,
in dependency edges).

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>