Fix a bug in array.find() for associative arrays indexed by strings
(and probably others as well).  Specifically, Engin noted that the
previous version assumed that it was iterating through indices in
order, but this can't be assumed to be the case for associative arrays
(which I believe are the only case that can fall into the serial case
at present).  Here, I took the naive approach of iterating through the
whole array, updating the index if it's lower than the previous one.

For a test, I created an associative array with lots of repeated
values since I couldn't predict what order the indices would be
visited in the current, or a future, implementation.  This failed
prior to this commit, but passes now, and I'm hoping is large enough
to have a high chance of failing should the associative array
implementation / iteration order change in the future.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>