Fix a bug in my parallel array.find()
I'd forgotten to write it so that the location where the index is
found is only updated if it's less than any previous found location.
This both handles cases where we might not be iterating over indices
in a well-defined order, and those common ones where we iterate over
them in increasing order and a single task has multiple matches. Duh.

Filed a test that didn't work prior to this fix and does now.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>