Add version of _format_reader for fileWriter and adjust both versions
Realized _format_reader was used by both reading and writing channels so needed
to support it with fileWriter.  While doing that, remove the parts that were
specialized for readf and writef calls when associated with the alternate
channel type.  I think this is right but I really hope we have good tests to
verify it o.o

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>