Compute dependencies when using --savec
For the C compiler, this computes dependencies via #includes when
compiling the generated .c file(s) the first time, such that any
subsequent edits to them by developers will re-compile the appropriate
files as necessary.

The approach taken here is to throw our standard DEPEND_CFLAGS when
doing a '--savec' compile, and to include the generated .d files
within the Makefile in such cases.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>